--- a/marshal.go
+++ b/marshal.go
@@ -84,15 +84,15 @@ func marshalStruct(val reflect.Value, info *reflectutils.SStructFieldInfo) JSONO
 }
 
 func struct2JSONPairs(val reflect.Value) []JSONPair {
-	objPairs := make([]JSONPair, 0)
 	fields := reflectutils.FetchStructFieldValueSet(val)
+	objPairs := make([]JSONPair, 0, len(fields))
 	for i := 0; i < len(fields); i += 1 {
 		jsonInfo := fields[i].Info
 		if jsonInfo.Ignore {
 			continue
 		}
 		key := jsonInfo.MarshalName()
-		val := marshalValue(fields[i].Value, &jsonInfo)
+		val := marshalValue(fields[i].Value, jsonInfo)
 		if val != nil && val != JSONNull {
 			objPair := JSONPair{key: key, val: val}
 			objPairs = append(objPairs, objPair)