@rainzm 2019-12-16
> jsonutils 不用包含 vendor

好的，那我先删了vendor，然后等 [PR](https://github.com/yunionio/pkg/pull/43) 合了再合这个。